re-organize remote e2e test to be pluggable
