+++ test/e2e_node/remote/types.go
@@ -19,6 +19,7 @@ limitations under the License.
 package remote
 
 import (
+	"fmt"
 	"time"
 )
 
@@ -50,4 +51,26 @@ type TestSuite interface {
 	// * runtimeConfig is the API runtime configuration used for node e2e tests.
 	// * timeout is the test timeout.
 	RunTest(host, workspace, results, imageDesc, junitFilePrefix, testArgs, ginkgoArgs, systemSpecName, extraEnvs, runtimeConfig string, timeout time.Duration) (string, error)
+}
+
+var testSuites = make(map[string]TestSuite)
+
+func RegisterTestSuite(name string, suite TestSuite) {
+	testSuites[name] = suite
+}
+
+func GetTestSuiteKeys() []string {
+	var keys []string
+	for key := range testSuites {
+		keys = append(keys, key)
+	}
+	return keys
+}
+
+func GetTestSuite(name string) (TestSuite, error) {
+	suite, ok := testSuites[name]
+	if ok {
+		return suite, nil
+	}
+	return nil, fmt.Errorf("unable to find testsuite for %s", name)
 }
